+++ project_cylon/BehaveController.py
@@ -1,7 +1,9 @@
 # -*- coding: utf-8 -*-
 from Logger import *
-from PageFactory import *
+from Responsive import *
+
 from CommonSteps import *
+from PageFactory import *
 
 from World import World as world
 
@@ -16,10 +18,11 @@ class BehaveController:
         if context.config.screenshot == True:
             Logger.enable_screenshots()
 
+        Responsive.load_browser_sizes("./responsive.yaml")
+
         site = "default"
         if hasattr(context.config, "site") and context.config.site is not None:
             site = context.config.site
-
         domain = PageFactory.get_domain("./config.yaml", site)
 
         pageobjects = "./pageobjects/*.yaml"
@@ -34,11 +37,16 @@ class BehaveController:
 
     @classmethod
     def before_feature(cls, context, feature):
+        if hasattr(context.config, "browser_size") and context.config.browser_size is not None:
+            name = context.config.browser_size
+            size = Responsive.get_browser_size(name)
+            world.size = size
+
         browser = "firefox"
         if hasattr(context.config, "browser") and context.config.browser is not None:
             browser = context.config.browser
-        world.open_browser(browser)
 
+        world.open_browser(browser)
         Logger.driver = world.driver
 
     @classmethod