+++ project_cylon/behook.py
@@ -5,6 +5,7 @@ from behave import __main__
 
 ## add my wanted option to parser.
 configuration.parser.add_argument('--browser')
+configuration.parser.add_argument('--browser-size')
 configuration.parser.add_argument('--site')
 configuration.parser.add_argument('--debug', action='store_true')
 configuration.parser.add_argument('--screenshot', action='store_true')